+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Use Active Record connection lock in `Database#synchronize` (@janko)
+
 ## 1.0.0 (2020-10-25)
 
 * Clear AR statement cache on `ActiveRecord::PreparedStatementCacheExpired` when Sequel holds the transaction (@janko)